refactor(frontend): Create centered layout for forms like auth
