+++ frontend/app/app.vue
@@ -1,5 +1,5 @@
 <template>
   <NuxtLayout>
-    <NuxtPage class="min-h-[calc(100vh-var(--ui-header-height))]" />
+    <NuxtPage />
   </NuxtLayout>
 </template>
