+++ frontend/app/layouts/centered.vue
@@ -1,0 +1,14 @@
+<script setup lang="ts">
+</script>
+
+<template>
+  <NuxtLayout name="default">
+    <div class="min-h-[calc(100vh-var(--ui-header-height))]">
+      <div class="flex flex-col md:flex-row justify-center items-center bg-gray-100 dark:bg-zinc-900 rendering-pixelated bg-[url(assets/img/background.webp)] bg-cover bg-center min-h-[calc(100vh-var(--ui-header-height))]">
+        <div class="space-y-4 w-90 bg-default p-8 rounded-2xl">
+          <slot />
+        </div>
+      </div>
+    </div>
+  </NuxtLayout>
+</template>
