+++ frontend/app/pages/request_password_reset.vue
@@ -5,6 +5,7 @@ import * as z from 'zod'
 
 definePageMeta({
   middleware: 'guest',
+  layout: 'centered',
 })
 
 type Schema = z.output<typeof schema>
@@ -50,7 +51,7 @@ async function onSubmit(event: FormSubmitEvent<Schema>) {
 </script>
 
 <template>
-  <AuthForm>
+  <div>
     <h1 class="text-2xl font-medium">
       Zresetuj hasło
     </h1>
@@ -75,5 +76,5 @@ async function onSubmit(event: FormSubmitEvent<Schema>) {
         </UButton>
       </div>
     </UForm>
-  </AuthForm>
+  </div>
 </template>
